Use transaction list items that change units alongside the current setting.
Connected to https://github.com/bithyve/hexa/issues/2368
