+++ src/components/account-details/AccountDetailsTransactionsListItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ListItem } from 'react-native-elements'
-import TransactionListItemContent from './TransactionListItemContent'
 import TransactionDescribing from '../../common/data/models/Transactions/Interfaces'
+import TransactionsFlatListItemContent from '../transactions/TransactionsFlatListItemContent'
 
 export type Props = {
   transaction: TransactionDescribing;
@@ -10,7 +10,7 @@ export type Props = {
 const AccountDetailsTransactionsListItem: React.FC<Props> = ( { transaction, }: Props ) => {
   return (
     <ListItem bottomDivider pad={4}>
-      <TransactionListItemContent transaction={transaction} />
+      <TransactionsFlatListItemContent transaction={transaction} />
       <ListItem.Chevron />
     </ListItem>
   )
