+++ src/components/account-settings/transaction-reassignment/TransactionsList.tsx
@@ -2,43 +2,43 @@ import React, { useCallback } from 'react'
 import { StyleSheet, FlatList } from 'react-native'
 import { ListItem } from 'react-native-elements'
 import RadioButton from '../../RadioButton'
-import TransactionListItemContent from '../../account-details/TransactionListItemContent'
 import TransactionDescribing from '../../../common/data/models/Transactions/Interfaces'
 import BitcoinUnit from '../../../common/data/enums/BitcoinUnit'
 import CurrencyKind from '../../../common/data/enums/CurrencyKind'
 import useCurrencyKind from '../../../utils/hooks/state-selectors/UseCurrencyKind'
+import TransactionListItemContent from '../../transactions/TransactionsFlatListItemContent'
 
 export type Props = {
   bitcoinUnit?: BitcoinUnit;
   currencyKind?: CurrencyKind | null;
   selectableTransactions: TransactionDescribing[];
   selectedTransactionIDs: Set<string>;
-  onTransactionSelected: (string) => void;
+  onTransactionSelected: ( string ) => void;
 };
 
-const listItemKeyExtractor = (item: TransactionDescribing) => item.txID
+const listItemKeyExtractor = ( item: TransactionDescribing ) => item.txID
 
-const TransactionsList: React.FC<Props> = ({
+const TransactionsList: React.FC<Props> = ( {
   bitcoinUnit = BitcoinUnit.SATS,
   currencyKind: defaultCurrencyKind,
   selectableTransactions,
   selectedTransactionIDs,
   onTransactionSelected,
-}: Props) => {
+}: Props ) => {
   const currencyKind = defaultCurrencyKind || useCurrencyKind()
 
-  const isChecked = useCallback((transaction: TransactionDescribing) => {
-    return selectedTransactionIDs.has(transaction.txID)
-  }, [ selectedTransactionIDs ])
+  const isChecked = useCallback( ( transaction: TransactionDescribing ) => {
+    return selectedTransactionIDs.has( transaction.txID )
+  }, [ selectedTransactionIDs ] )
 
-  const renderItem = ({ item: transaction }: { item: TransactionDescribing }) => {
+  const renderItem = ( { item: transaction }: { item: TransactionDescribing } ) => {
     return (
       <ListItem
         bottomDivider
-        onPress={() => { onTransactionSelected(transaction) }}
+        onPress={() => { onTransactionSelected( transaction ) }}
       >
         <RadioButton
-          isChecked={isChecked(transaction)}
+          isChecked={isChecked( transaction )}
           ignoresTouch
         />
 
@@ -55,7 +55,7 @@ const TransactionsList: React.FC<Props> = ({
     <FlatList
       style={styles.rootContainer}
       contentContainerStyle={{
-        paddingHorizontal: 14 
+        paddingHorizontal: 14
       }}
       extraData={selectedTransactionIDs}
       data={selectableTransactions}
@@ -65,9 +65,9 @@ const TransactionsList: React.FC<Props> = ({
   )
 }
 
-const styles = StyleSheet.create({
+const styles = StyleSheet.create( {
   rootContainer: {
   }
-})
+} )
 
 export default TransactionsList
