+++ src/components/transactions/TransactionsFlatListItem.tsx
@@ -1,22 +1,23 @@
-import React from 'react';
-import { ListItem } from 'react-native-elements';
-import TransactionsFlatListItemContent from './TransactionsFlatListItemContent';
-import { TransactionDetails } from '../../bitcoin/utilities/Interface';
+import React from 'react'
+import { ListItem } from 'react-native-elements'
+import TransactionsFlatListItemContent from './TransactionsFlatListItemContent'
+import { TransactionDetails } from '../../bitcoin/utilities/Interface'
 
 export type Props = {
   transaction: TransactionDetails;
   onPress: () => void;
 };
 
-const TransactionsFlatListItem: React.FC<Props> = ({
+const TransactionsFlatListItem: React.FC<Props> = ( {
   transaction,
   onPress,
-}: Props) => {
+}: Props ) => {
   return (
     <ListItem bottomDivider onPress={onPress}>
       <TransactionsFlatListItemContent transaction={transaction} />
+      <ListItem.Chevron />
     </ListItem>
-  );
-};
+  )
+}
 
-export default TransactionsFlatListItem;
+export default TransactionsFlatListItem
